rewrite LALMServing (#219)
